Address all places that relied on integer primary key

+++ scripts/migrations/20170407022844409_create_profile.cr
@@ -1,7 +1,7 @@
 class CreateProfile20170407022844409 < Jennifer::Migration::Base
   def up
     create_table(:profiles) do |t|
-      t.integer :contact_id
+      t.bigint :contact_id
       t.string :type, { :null => false }
       t.string :uid
       t.string :login
